cpufreq_sublime: make frequency changes proportional to a boundary frequency
Any increases or decreases will be proportional to the difference
between the current frequency and boundary frequency. For example if the
frequency is being increased then the frequency change would be
proportional to the current frequency and the maximum frequency. The
same is true for frequency decreases. This should make changes in
frequency responsive while reducing the amount of frequency increases
primarily at higher frequencies. Furthermore, the micro frequency step is not needed anymore since boundary frequencies can be used to tune how much the frequency increases or decreases.

fixes for prop gov redo

Edit default values