input/touchboost.c: only calculate the current time once
When checking to see if the last input event is sufficiently far enough
from the current input event, the current time is calculated
twice. Calculate the time once using the CURRENT_TIME_US macro and store
it in now. If the last input event is far enough from the current time,
then set the last input event to now instead of recalculating the current
time.