cpufreq_sublime: Ensure a minimum frequency change
The maximum base frequency delta is obsolete since frequency the changes
have a guaranteed minimum value. Additionally, the base frequency delta
leads to the frequency going over the maximum frequency or under the
minimum frequency. With this new implementation, the frequency to
guaranteed to stay within the acceptable boundaries set by each CPU.