tegra-fuse: fix fuse_cp_rev_check()
There is an issue with the fuse_cp_rev_check() not being a static
function since it had a static variable. If it is to be called in
denver_knobs.c then it cannot have any static variables. Furthermore, it
must be included in a header file for compatibility with gcc 5 and
up. Lastly, this appears to fix an issue with charging the
battery. Without this commit, the voltage on the device is not set
properly, and consequently the battery level does not update properly
when charging.