cpufreq_sublime: Be more responsive to input events
When an input event is made, try to keep the frequency at the minimum
input event frequency. This ensures a responsive experience while
maintaining minimizing battery use. However if the load is high, then
increase the minimum input event frequency by an amount proportional to
load. This makes the CPU more responsive to higher loads.