cpufreq_sublimeactive: Simplify how frequency is set during input event.
During input event, the CPU frequency was scaled proportional to the
load. Instead of making a special case for scaling the frequency for an
input event, the frequency will be guaranteed to be at least the minimum
input event frequency in the general case. This makes the code easier to
understand and performance will be about the same.