cpufreq_sublimeaction: Remove boundary checks.
Previously a check is made to ensure the CPU frequency is at least
policy->min and at most policy->max. However, with the way the CPU
frequency is scaled, it is impossible for it to go outside the
boundaries. If the frequency is at its maximum value, then it would be
doubled then divided by 2. The same is true for the when the frequency
is at its minimum value.