cpufreq/sa: Adjust cpufreq relations
Previously, thr relations were set so that middle frequencies would be favored. For example, if increasing the frequency, the current and max frequencies would be averaged and rounded downwards to the nearest frequency. Likewise when reducing the frequency, The current and minimum frequencies are averaged and rounded to the next highest frequency. I see this as being useful only if the boundaries can be violated. However __cpufreq_driver_target() already does a boundary checks. Making this change will make scaling the cpu frequency more responsive.