cpuquiet: touch: Schedule all work on CPU 0
Previously, the work for checking a CPU could be done on the CPU
that is checked. Yet, only CPU 0 is consistently online. This could cause
poor performance if the CPU checking work is run on a different CPU each time.
Running all work on CPU 0 increases performance due to cache locality.

The dedicated workqueue is no longer needed as a result of this change since it
only needs the WQ_FREEZABLE flag. The system_freezeable_wq has this flag so lets
use it instead.